src/requirements.txt
numpy
opencv-python==3.4.2.17
pyyaml
tqdm
matplotlib